requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
langchain-core
pydantic
playwright
python-dotenv
requests
langgraph-checkpoint-sqlite
aiosqlite
langchain-openai
langchain-anthropic
langchain-aws
boto3
